Adding consolte test addIn program.
Refactor and enhance library interface and error handling

- Changed return type of `GetClassObject` to `intptr_t` in `ComponentBase.h` and `common_main.cpp` for better pointer handling.
- Updated preprocessor definitions in `ECRDriver.vcxproj` and removed `PUGIXML_WCHAR_MODE`.
- Improved error handling in multiple overloads of `loadValue` in `addin_driver.cpp`.
- Commented out export definitions in `common_main.h` to indicate potential changes in function exposure.
- Added parameter count checks in methods of `DriverPOSTerminal` class in `driver_pos_terminal.cpp`.
- Updated `exports.def` to include `GetClassObject` and related functions in the public interface.
- Modified `interface_pos_terminal.h` to include additional headers and removed the `CashWithdrawal` method.
- Enhanced OpenSSL error handling in `license_manager.cpp`.
- Defined `LIB_EXPORTS` and `U_STATIC_IMPLEMENTATION` in `pch.h` for precompiled header management.
- Introduced new JSON utility functions and a TCP server in `emul_terminal.py`.
- Updated `test_addin.cpp` to include new tests for library functions.
- Adjusted project files to include new source files and updated include paths.

+++ include/vncomp/ComponentBase.h
@@ -256,7 +256,7 @@ enum AttachType
 /**
  * These functions should be implemented that component can be loaded and created. 
  */
-extern "C" long GetClassObject(const WCHAR_T*, IComponentBase** pIntf);
+extern "C" intptr_t GetClassObject(const WCHAR_T*, IComponentBase** pIntf);
 extern "C" long DestroyObject(IComponentBase** pIntf);
 extern "C" const WCHAR_T* GetClassNames();
 extern "C" AppCapabilities SetPlatformCapabilities(const AppCapabilities capabilities);
